+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this project are documented below.
 The format is based on [keep a changelog](http://keepachangelog.com) and this project uses [semantic versioning](http://semver.org).
 
 ## [Unreleased]
-
+### Added
+- Add subscription validation APIs and runtime functions for Google and Apple.
 
 ## [1.23.0] - 2022-05-22
 ### Added
@@ -18,7 +19,7 @@ The format is based on [keep a changelog](http://keepachangelog.com) and this pr
 - Update naming of delete notifications before/after hook registration functions.
 - Allow storage writes to accept "undefined" as a user ID input in the TypeScript definitions.
 
-### Fixed
+### Fixed
 - Rename field in "ValidatedPurchase" to "providerResponse" to match what the server returns.
 - Add missing "label" field to "Match" TypeScript definition.
 
@@ -41,7 +42,7 @@ The format is based on [keep a changelog](http://keepachangelog.com) and this pr
 
 ### Changed
 - JavaScript runtime match data changed to use Uint8Array type.
-- Added IAP validation `persist` parameter to make persistence optional. 
+- Added IAP validation `persist` parameter to make persistence optional.
 
 ## [1.20.0] - 2021-10-29
 ### Added